Day 3 stuff

+++ Week01/Day2Assignments/assignment3oconnell.c
@@ -4,18 +4,18 @@
 
 int main()                /* a program that gets three numbers and does prism computations with three numbers */
 {
-	float height, depth, width;
-	float sat, saff, sae, sap; // Surface area of top, of front face, of end, and the whole prism
-	float v; // Volume
-	float pb, pff, pe; // Perimeter bottom, front face, and end
+	int height, depth, width;
+	int sat, saff, sae, sap; // Surface area of top, of front face, of end, and the whole prism
+	int v; // Volume
+	int pb, pff, pe; // Perimeter bottom, front face, and end
 	float d; // Diagonal
 	
 	printf("What is the height?\n");
-	scanf("%f", &height);
+	scanf("%d", &height);
 	printf("What is the depth?\n");
-	scanf("%f", &depth);
+	scanf("%d", &depth);
 	printf("What is the height?\n");
-	scanf("%f", &width);
+	scanf("%d", &width);
 	
 	sat = depth * width; // Calculates Surface area of top
 	saff = height * width; // of front face
@@ -23,24 +23,24 @@ int main()                /* a program that gets three numbers and does prism co
 	sap = 2 * (sat + saff + sae); // and the whole prism
 	
     printf("\nSurface Areas: \n"); // Prints the surface areas all nice
-    printf("Top: %.2f,",sat);
-    printf(" Front face: %.2f,",saff);
-    printf(" End: %.2f,",sae);
-    printf(" Whole Prism: %.2f\n",sap);  
+    printf("Top: %d,",sat);
+    printf(" Front face: %d,",saff);
+    printf(" End: %d,",sae);
+    printf(" Whole Prism: %d\n",sap);  
     
     pb = 2 * (depth + width); // Perimeter of bottom
     pff = 2 * (height + width); // front face
     pe = 2 * (height + depth); // and end
     
     printf("\nPerimeters: \n"); // Prints the perimeters all nice
-    printf("Bottom: %.2f,",pb);
-    printf(" Front face: %.2f,",pff);
-    printf(" End: %.2f\n",pe);
+    printf("Bottom: %d,",pb);
+    printf(" Front face: %d,",pff);
+    printf(" End: %d\n",pe);
     
     v = height * depth * width; // Volumes
-    printf("\nVolume: %.2f \n", v); // Prints the volume
+    printf("\nVolume: %d \n", v); // Prints the volume
     
-    d = sqrtf(depth*depth + height*height + width*width); // Diagonal
+    d = sqrtf((float)depth*depth + height*height + width*width); // Diagonal
     printf("\nDiagonal through the center of the prism: %.2f\n",d); // Prints the diagonal
     
     return 0;
